+++ apps/backend/src/modules/friends/friend.repository.ts
@@ -1,6 +1,14 @@
+import { type Page } from "objection";
+
+import { SortOrder } from "~/libs/enums/enums.js";
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
-import { type Repository } from "~/libs/types/types.js";
+import {
+	type PaginationResponseDto,
+	type Repository,
+} from "~/libs/types/types.js";
 import { EMPTY_ARRAY_LENGTH } from "~/libs/types/types.js";
+import { GroupEntity } from "~/modules/groups/group.entity.js";
+import { PermissionEntity } from "~/modules/permissions/permissions.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 import { type UserModel } from "~/modules/users/user.model.js";
 
@@ -29,7 +37,7 @@ class FriendRepository implements Repository<UserEntity> {
 			.query()
 			.findById(followingUserId)
 			.withGraphJoined(
-				`${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
+				`[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
 			)
 			.castTo<UserModel>();
 
@@ -38,6 +46,24 @@ class FriendRepository implements Repository<UserEntity> {
 			createdAt: followingUser.createdAt,
 			email: followingUser.email,
 			firstName: followingUser.userDetails.firstName,
+			groups: followingUser.groups.map((group) => {
+				return GroupEntity.initialize({
+					createdAt: group.createdAt,
+					id: group.id,
+					key: group.key,
+					name: group.name,
+					permissions: group.permissions.map((permission) => {
+						return PermissionEntity.initialize({
+							createdAt: permission.createdAt,
+							id: permission.id,
+							key: permission.key,
+							name: permission.name,
+							updatedAt: permission.updatedAt,
+						});
+					}),
+					updatedAt: group.updatedAt,
+				});
+			}),
 			id: followingUser.id,
 			lastName: followingUser.userDetails.lastName,
 			nickname: followingUser.userDetails.nickname,
@@ -85,7 +111,7 @@ class FriendRepository implements Repository<UserEntity> {
 			.query()
 			.findById(id)
 			.withGraphJoined(
-				`${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
+				`[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
 			)
 			.execute();
 
@@ -95,6 +121,24 @@ class FriendRepository implements Repository<UserEntity> {
 					createdAt: user.createdAt,
 					email: user.email,
 					firstName: user.userDetails.firstName,
+					groups: user.groups.map((group) => {
+						return GroupEntity.initialize({
+							createdAt: group.createdAt,
+							id: group.id,
+							key: group.key,
+							name: group.name,
+							permissions: group.permissions.map((permission) => {
+								return PermissionEntity.initialize({
+									createdAt: permission.createdAt,
+									id: permission.id,
+									key: permission.key,
+									name: permission.name,
+									updatedAt: permission.updatedAt,
+								});
+							}),
+							updatedAt: group.updatedAt,
+						});
+					}),
 					id: user.id,
 					lastName: user.userDetails.lastName,
 					nickname: user.userDetails.nickname,
@@ -116,7 +160,7 @@ class FriendRepository implements Repository<UserEntity> {
 				`${DatabaseTableName.FRIENDS}.following_id`,
 			)
 			.withGraphJoined(
-				`${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
+				`[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
 			);
 
 		return followings.map((user) => {
@@ -125,6 +169,24 @@ class FriendRepository implements Repository<UserEntity> {
 				createdAt: user.createdAt,
 				email: user.email,
 				firstName: user.userDetails.firstName,
+				groups: user.groups.map((group) => {
+					return GroupEntity.initialize({
+						createdAt: group.createdAt,
+						id: group.id,
+						key: group.key,
+						name: group.name,
+						permissions: group.permissions.map((permission) => {
+							return PermissionEntity.initialize({
+								createdAt: permission.createdAt,
+								id: permission.id,
+								key: permission.key,
+								name: permission.name,
+								updatedAt: permission.updatedAt,
+							});
+						}),
+						updatedAt: group.updatedAt,
+					});
+				}),
 				id: user.id,
 				lastName: user.userDetails.lastName,
 				nickname: user.userDetails.nickname,
@@ -171,8 +233,16 @@ class FriendRepository implements Repository<UserEntity> {
 		return Boolean(subscription);
 	}
 
-	public async getPotentialFollowers(id: number): Promise<UserEntity[]> {
-		const potentialFollowers = await this.userModel
+	public async getPotentialFollowers({
+		count,
+		id,
+		page,
+	}: {
+		count: number;
+		id: number;
+		page: number;
+	}): Promise<PaginationResponseDto<UserEntity>> {
+		const { results, total } = await this.userModel
 			.query()
 			.where("users.id", "<>", id)
 			.leftJoin(
@@ -197,28 +267,60 @@ class FriendRepository implements Repository<UserEntity> {
 			)
 			.distinct()
 			.withGraphJoined(
-				`${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
-			);
+				`[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
+			)
+			.orderBy(`${DatabaseTableName.USERS}.createdAt`, SortOrder.DESC)
+			.page(page, count)
+			.castTo<Page<UserModel>>();
 
-		return potentialFollowers.map((user) => {
-			return UserEntity.initialize({
-				avatarUrl: user.userDetails.avatarFile?.url ?? null,
-				createdAt: user.createdAt,
-				email: user.email,
-				firstName: user.userDetails.firstName,
-				id: user.id,
-				lastName: user.userDetails.lastName,
-				nickname: user.userDetails.nickname,
-				passwordHash: user.passwordHash,
-				passwordSalt: user.passwordSalt,
-				sex: user.userDetails.sex,
-				updatedAt: user.updatedAt,
-			});
-		});
+		return {
+			items: results.map((user) => {
+				return UserEntity.initialize({
+					avatarUrl: user.userDetails.avatarFile?.url ?? null,
+					createdAt: user.createdAt,
+					email: user.email,
+					firstName: user.userDetails.firstName,
+					groups: user.groups.map((group) => {
+						return GroupEntity.initialize({
+							createdAt: group.createdAt,
+							id: group.id,
+							key: group.key,
+							name: group.name,
+							permissions: group.permissions.map((permission) => {
+								return PermissionEntity.initialize({
+									createdAt: permission.createdAt,
+									id: permission.id,
+									key: permission.key,
+									name: permission.name,
+									updatedAt: permission.updatedAt,
+								});
+							}),
+							updatedAt: group.updatedAt,
+						});
+					}),
+					id: user.id,
+					lastName: user.userDetails.lastName,
+					nickname: user.userDetails.nickname,
+					passwordHash: user.passwordHash,
+					passwordSalt: user.passwordSalt,
+					sex: user.userDetails.sex,
+					updatedAt: user.updatedAt,
+				});
+			}),
+			total,
+		};
 	}
 
-	public async getUserFollowers(id: number): Promise<UserEntity[]> {
-		const userFollowers = await this.userModel
+	public async getUserFollowers({
+		count,
+		id,
+		page,
+	}: {
+		count: number;
+		id: number;
+		page: number;
+	}): Promise<PaginationResponseDto<UserEntity>> {
+		const { results, total } = await this.userModel
 			.query()
 			.leftJoin(
 				DatabaseTableName.FRIENDS,
@@ -228,28 +330,60 @@ class FriendRepository implements Repository<UserEntity> {
 			)
 			.where(`${DatabaseTableName.FRIENDS}.following_id`, "=", id)
 			.withGraphJoined(
-				`${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
-			);
+				`[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
+			)
+			.orderBy(`${DatabaseTableName.FRIENDS}.updatedAt`, SortOrder.DESC)
+			.page(page, count)
+			.castTo<Page<UserModel>>();
 
-		return userFollowers.map((user) => {
-			return UserEntity.initialize({
-				avatarUrl: user.userDetails.avatarFile?.url ?? null,
-				createdAt: user.createdAt,
-				email: user.email,
-				firstName: user.userDetails.firstName,
-				id: user.id,
-				lastName: user.userDetails.lastName,
-				nickname: user.userDetails.nickname,
-				passwordHash: user.passwordHash,
-				passwordSalt: user.passwordSalt,
-				sex: user.userDetails.sex,
-				updatedAt: user.updatedAt,
-			});
-		});
+		return {
+			items: results.map((user) => {
+				return UserEntity.initialize({
+					avatarUrl: user.userDetails.avatarFile?.url ?? null,
+					createdAt: user.createdAt,
+					email: user.email,
+					firstName: user.userDetails.firstName,
+					groups: user.groups.map((group) => {
+						return GroupEntity.initialize({
+							createdAt: group.createdAt,
+							id: group.id,
+							key: group.key,
+							name: group.name,
+							permissions: group.permissions.map((permission) => {
+								return PermissionEntity.initialize({
+									createdAt: permission.createdAt,
+									id: permission.id,
+									key: permission.key,
+									name: permission.name,
+									updatedAt: permission.updatedAt,
+								});
+							}),
+							updatedAt: group.updatedAt,
+						});
+					}),
+					id: user.id,
+					lastName: user.userDetails.lastName,
+					nickname: user.userDetails.nickname,
+					passwordHash: user.passwordHash,
+					passwordSalt: user.passwordSalt,
+					sex: user.userDetails.sex,
+					updatedAt: user.updatedAt,
+				});
+			}),
+			total,
+		};
 	}
 
-	public async getUserFollowings(id: number): Promise<UserEntity[]> {
-		const userFollowings = await this.userModel
+	public async getUserFollowings({
+		count,
+		id,
+		page,
+	}: {
+		count: number;
+		id: number;
+		page: number;
+	}): Promise<PaginationResponseDto<UserEntity>> {
+		const { results, total } = await this.userModel
 			.query()
 			.leftJoin(
 				DatabaseTableName.FRIENDS,
@@ -259,24 +393,48 @@ class FriendRepository implements Repository<UserEntity> {
 			)
 			.where(`${DatabaseTableName.FRIENDS}.follower_id`, "=", id)
 			.withGraphJoined(
-				`${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
-			);
+				`[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
+			)
+			.orderBy(`${DatabaseTableName.FRIENDS}.updatedAt`, SortOrder.DESC)
+			.page(page, count)
+			.castTo<Page<UserModel>>();
 
-		return userFollowings.map((user) => {
-			return UserEntity.initialize({
-				avatarUrl: user.userDetails.avatarFile?.url ?? null,
-				createdAt: user.createdAt,
-				email: user.email,
-				firstName: user.userDetails.firstName,
-				id: user.id,
-				lastName: user.userDetails.lastName,
-				nickname: user.userDetails.nickname,
-				passwordHash: user.passwordHash,
-				passwordSalt: user.passwordSalt,
-				sex: user.userDetails.sex,
-				updatedAt: user.updatedAt,
-			});
-		});
+		return {
+			items: results.map((user) => {
+				return UserEntity.initialize({
+					avatarUrl: user.userDetails.avatarFile?.url ?? null,
+					createdAt: user.createdAt,
+					email: user.email,
+					firstName: user.userDetails.firstName,
+					groups: user.groups.map((group) => {
+						return GroupEntity.initialize({
+							createdAt: group.createdAt,
+							id: group.id,
+							key: group.key,
+							name: group.name,
+							permissions: group.permissions.map((permission) => {
+								return PermissionEntity.initialize({
+									createdAt: permission.createdAt,
+									id: permission.id,
+									key: permission.key,
+									name: permission.name,
+									updatedAt: permission.updatedAt,
+								});
+							}),
+							updatedAt: group.updatedAt,
+						});
+					}),
+					id: user.id,
+					lastName: user.userDetails.lastName,
+					nickname: user.userDetails.nickname,
+					passwordHash: user.passwordHash,
+					passwordSalt: user.passwordSalt,
+					sex: user.userDetails.sex,
+					updatedAt: user.updatedAt,
+				});
+			}),
+			total,
+		};
 	}
 
 	public async update(id: number): Promise<UserEntity> {
@@ -292,6 +450,24 @@ class FriendRepository implements Repository<UserEntity> {
 			createdAt: updatedSubscription.createdAt,
 			email: updatedSubscription.email,
 			firstName: updatedSubscription.userDetails.firstName,
+			groups: updatedSubscription.groups.map((group) => {
+				return GroupEntity.initialize({
+					createdAt: group.createdAt,
+					id: group.id,
+					key: group.key,
+					name: group.name,
+					permissions: group.permissions.map((permission) => {
+						return PermissionEntity.initialize({
+							createdAt: permission.createdAt,
+							id: permission.id,
+							key: permission.key,
+							name: permission.name,
+							updatedAt: permission.updatedAt,
+						});
+					}),
+					updatedAt: group.updatedAt,
+				});
+			}),
 			id: updatedSubscription.id,
 			lastName: updatedSubscription.userDetails.lastName,
 			nickname: updatedSubscription.userDetails.nickname,