+++ apps/backend/src/modules/groups/libs/enums/relation-name.enum.ts
@@ -1,0 +1,9 @@
+const RelationName = {
+	AVATAR_FILE: "avatarFile",
+	GROUPS: "groups",
+	PERMISSIONS: "permissions",
+	USER_DETAILS: "userDetails",
+	USERS: "users",
+} as const;
+
+export { RelationName };