+++ packages/shared/src/libs/helpers/check-is-today/check-is-today.helper.ts
@@ -1,0 +1,7 @@
+import { isToday } from "date-fns";
+
+const checkIsToday = (date: Date | number | string): boolean => {
+	return isToday(date);
+};
+
+export { checkIsToday };