+++ packages/shared/src/modules/permissions/libs/enums/permissions-api-path.enum.ts
@@ -1,0 +1,6 @@
+const PermissionsApiPath = {
+	$PERMISSION_ID: "/:permissionId",
+	ROOT: "/",
+} as const;
+
+export { PermissionsApiPath };